LPS-60835 Move simulation control menu entry to simulation portlet
