+++ modules/apps/control-menu/control-menu-web/src/main/java/com/liferay/control/menu/web/portlet/ControlMenuPortlet.java
@@ -14,9 +14,6 @@
 
 package com.liferay.control.menu.web.portlet;
 
-import com.liferay.application.list.PanelAppRegistry;
-import com.liferay.application.list.PanelCategoryRegistry;
-import com.liferay.application.list.constants.ApplicationListWebKeys;
 import com.liferay.control.menu.ControlMenuCategory;
 import com.liferay.control.menu.constants.ControlMenuCategoryKeys;
 import com.liferay.control.menu.constants.ControlMenuWebKeys;
@@ -127,11 +124,6 @@ public class ControlMenuPortlet extends MVCPortlet {
 		renderRequest.setAttribute(
 			ControlMenuWebKeys.CONTROL_MENU_ENTRY_REGISTRY,
 			_controlMenuEntryRegistry);
-		renderRequest.setAttribute(
-			ApplicationListWebKeys.PANEL_APP_REGISTRY, _panelAppRegistry);
-		renderRequest.setAttribute(
-			ApplicationListWebKeys.PANEL_CATEGORY_REGISTRY,
-			_panelCategoryRegistry);
 
 		super.doDispatch(renderRequest, renderResponse);
 	}
@@ -161,21 +153,7 @@ public class ControlMenuPortlet extends MVCPortlet {
 		_controlMenuEntryRegistry = controlMenuEntryRegistry;
 	}
 
-	@Reference(unbind = "-")
-	protected void setPanelAppRegistry(PanelAppRegistry panelAppRegistry) {
-		_panelAppRegistry = panelAppRegistry;
-	}
-
-	@Reference(unbind = "-")
-	protected void setPanelCategoryRegistry(
-		PanelCategoryRegistry panelCategoryRegistry) {
-
-		_panelCategoryRegistry = panelCategoryRegistry;
-	}
-
 	private volatile ControlMenuCategoryRegistry _controlMenuCategoryRegistry;
 	private volatile ControlMenuEntryRegistry _controlMenuEntryRegistry;
-	private volatile PanelAppRegistry _panelAppRegistry;
-	private volatile PanelCategoryRegistry _panelCategoryRegistry;
 
 }
