+++ modules/apps/control-menu/control-menu-web/src/main/resources/META-INF/resources/init.jsp
@@ -18,19 +18,14 @@
 
 <%@ taglib uri="http://java.sun.com/portlet_2_0" prefix="portlet" %>
 
-<%@ taglib uri="http://liferay.com/tld/application-list" prefix="liferay-application-list" %><%@
-taglib uri="http://liferay.com/tld/aui" prefix="aui" %><%@
+<%@ taglib uri="http://liferay.com/tld/aui" prefix="aui" %><%@
 taglib uri="http://liferay.com/tld/portlet" prefix="liferay-portlet" %><%@
 taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %><%@
 taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %><%@
 taglib uri="http://liferay.com/tld/util" prefix="liferay-util" %>
 
-<%@ page import="com.liferay.application.list.PanelCategory" %><%@
-page import="com.liferay.application.list.PanelCategoryRegistry" %><%@
-page import="com.liferay.application.list.constants.ApplicationListWebKeys" %><%@
-page import="com.liferay.control.menu.ControlMenuCategory" %><%@
+<%@ page import="com.liferay.control.menu.ControlMenuCategory" %><%@
 page import="com.liferay.control.menu.ControlMenuEntry" %><%@
-page import="com.liferay.control.menu.application.list.SimulationPanelCategory" %><%@
 page import="com.liferay.control.menu.constants.ControlMenuWebKeys" %><%@
 page import="com.liferay.control.menu.util.ControlMenuEntryRegistry" %><%@
 page import="com.liferay.control.menu.web.InformationMessagesControlMenuEntry" %><%@
