+++ modules/apps/product-navigation/product-navigation-simulation-web/build.gradle
@@ -1,6 +1,8 @@
 dependencies {
 	compile project(":apps:application-list:application-list-api")
 	compile project(":apps:application-list:application-list-taglib")
+	compile project(":apps:control-menu:control-menu-api")
+	compile project(":apps:product-navigation:product-navigation-simulation")
 }
 
 liferay {
