+++ modules/apps/product-navigation/product-navigation-simulation-web/src/main/resources/META-INF/resources/portlet/view.jsp
@@ -15,3 +15,11 @@
 --%>
 
 <%@ include file="/portlet/init.jsp" %>
+
+<%
+PanelCategoryRegistry panelCategoryRegistry = (PanelCategoryRegistry)request.getAttribute(ApplicationListWebKeys.PANEL_CATEGORY_REGISTRY);
+
+PanelCategory panelCategory = panelCategoryRegistry.getPanelCategory(SimulationPanelCategory.SIMULATION);
+%>
+
+<liferay-application-list:panel-category panelCategory="<%= panelCategory %>" showHeader="<%= false %>" />
