+++ portal-impl/src/portal.properties
@@ -553,7 +553,7 @@
     # Set a list of comma delimited portlet IDs that will bypass the security
     # check set in the property "portlet.add.default.resource.check.enabled".
     #
-    portlet.add.default.resource.check.whitelist=com_liferay_login_web_portlet_LoginPortlet,com_liferay_portlet_configuration_web_portlet_PortletConfigurationPortlet,com_liferay_portlet_css_web_portlet_PortletCSSPortlet,145,164,166,com_liferay_staging_bar_web_portlet_StagingBarPortlet
+    portlet.add.default.resource.check.whitelist=com_liferay_login_web_portlet_LoginPortlet,com_liferay_portlet_configuration_web_portlet_PortletConfigurationPortlet,com_liferay_portlet_css_web_portlet_PortletCSSPortlet,145,164,166,com_liferay_product_navigation_simulation_web_portlet_SimulationPortlet,com_liferay_staging_bar_web_portlet_StagingBarPortlet"
 
     #
     # Input a list of comma delimited struts actions that will bypass the
